+++ cx-portal/src/features/serviceManagement/apiSlice.ts
@@ -19,10 +19,7 @@
  ********************************************************************************/
 
 import { createApi, fetchBaseQuery } from '@reduxjs/toolkit/query/react'
-import {
-  DocumentRequestData,
-  DocumentTypeId,
-} from 'features/appManagement/apiSlice'
+import { DocumentTypeId } from 'features/appManagement/apiSlice'
 import { apiBaseQuery } from 'utils/rtkUtil'
 import { ServiceStatusDataState } from './types'
 
@@ -124,7 +121,7 @@ export const apiSlice = createApi({
       UpdateAgreementConsentType
     >({
       query: (data: UpdateAgreementConsentType) => ({
-        url: `/api/services/${data.appId}/serviceAgreementConsent`,
+        url: `/api/services/servicerelease/consent/${data.appId}/agreementConsents`,
         method: 'POST',
         body: data.body,
       }),
@@ -161,16 +158,7 @@ export const apiSlice = createApi({
     }),
     fetchNewDocumentById: builder.mutation({
       query: (documentId) => ({
-        url: `/api/administration/documents/${documentId}`,
-        responseHandler: async (response) => ({
-          headers: response.headers,
-          data: await response.blob(),
-        }),
-      }),
-    }),
-    fetchDocumentById: builder.mutation({
-      query: (data: DocumentRequestData) => ({
-        url: `/api/apps/${data.appId}/appDocuments/${data.documentId}`,
+        url: `/api/administration/documents/frameDocuments/${documentId}`,
         responseHandler: async (response) => ({
           headers: response.headers,
           data: await response.blob(),
@@ -190,5 +178,4 @@ export const {
   useFetchServiceConsentDataQuery,
   useUpdateServiceDocumentUploadMutation,
   useFetchNewDocumentByIdMutation,
-  useFetchDocumentByIdMutation,
 } = apiSlice